added first step of checkout and fixed bug in my account page
The first step of checkout was added, the one where the user see all it's cart's items and the price to which they all add up to. Also fixed the bug in my account page that made the application crash when the user reloaded there. Also, footer was fixed and now stays in it's place. I also tweaked a few design related things in the main page